Move nvc commands to subcommand
