+++ cmd/nvccmd/create.go
@@ -1,4 +1,4 @@
-package cmd
+package nvccmd
 
 import (
 	"errors"
